refactor: add inline comments and move regression tests into TestTransactionFormStructure
- Add detailed inline comment on the on_commit deferral in
  TransactionCreateForm.save() explaining why the change was needed
  and what would happen without it
- Move the two #333 regression tests from the standalone
  test_transaction_create_form_on_commit.py into the existing
  TestTransactionFormStructure class in test_transaction_form_structure.py,
  which is the natural home for TransactionCreateForm behaviour tests
- Delete the now-redundant standalone file
